# Conflicts:
#	out/production/GUI/Accounts/Accounts_Controller.class
#	out/production/GUI/Accounts/Administrator.class
#	out/production/GUI/Accounts/ForePerson.class
#	out/production/GUI/Accounts/Franchisee.class
#	out/production/GUI/Accounts/I_Accounts.class
#	out/production/GUI/Accounts/Mechanic.class
#	out/production/GUI/Accounts/Receptionist.class
#	out/production/GUI/Accounts/SQL_UserHelper.class
#	out/production/GUI/Accounts/User.class
#	out/production/GUI/Accounts/UserCollection.class
#	out/production/GUI/AdminLoginGUI$1.class
#	out/production/GUI/AdminLoginGUI$2.class
#	out/production/GUI/AdminLoginGUI.class
#	out/production/GUI/CreateJobGUI$1.class
#	out/production/GUI/CreateJobGUI$2.class
#	out/production/GUI/CreateJobGUI.class
#	out/production/GUI/Customers/Customer.class
#	out/production/GUI/Customers/SQL_CustomerHelper.class
#	out/production/GUI/Customers/SQL_VehicleHelper.class
#	out/production/GUI/Database/Database_Controller.class
#	out/production/GUI/ItemSearchGUI$1.class
#	out/production/GUI/ItemSearchGUI$2.class
#	out/production/GUI/ItemSearchGUI$3.class
#	out/production/GUI/ItemSearchGUI.class
#	out/production/GUI/Job/Job.class
#	out/production/GUI/Job/Job_Controller.class
#	out/production/GUI/Job/SQL_JobHelper.class
#	out/production/GUI/Job/SQL_PartsHelper.class
#	out/production/GUI/Job/SparePart.class
#	out/production/GUI/JobDetailsGUI$1.class
#	out/production/GUI/JobDetailsGUI.class
#	out/production/GUI/JobReportGUI$2.class
#	out/production/GUI/JobReportGUI.class
#	out/production/GUI/JobsMenuGUI.class
#	out/production/GUI/LoginGUI$1.class
#	out/production/GUI/LoginGUI.class
#	out/production/GUI/MainMenuGUI$3.class
#	out/production/GUI/MainMenuGUI$4.class
#	out/production/GUI/MainMenuGUI$5.class
#	out/production/GUI/MainMenuGUI.class
#	out/production/GUI/SearchResultsGUI$1.class
#	out/production/GUI/SearchResultsGUI.class
#	out/production/GUI/SplashScreen.class
#	out/production/GUI/StockReportGUI$2.class
#	out/production/GUI/StockReportGUI.class
#	out/production/GUI/VehicleReportGUI$2.class
#	out/production/GUI/VehicleReportGUI.class
#	out/production/GUI/ViewJobsGUI$1.class
#	out/production/GUI/ViewJobsGUI$2.class
#	out/production/GUI/ViewJobsGUI.class
#	out/production/GUI/main.class

+++ src/Job/MOT.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class MOT extends Job {
 
 	private boolean pass;
-	private long mileage;
+	private int mileage;
 	private long MOTTestNumber;
 	private Date expiryDate;
 	private String testLocation;
@@ -24,7 +24,7 @@ public class MOT extends Job {
 		this.pass = pass;
 	}
 
-	public long getMileage() {
+	public int getMileage() {
 		return this.mileage;
 	}
 
@@ -32,7 +32,7 @@ public class MOT extends Job {
 	 * 
 	 * @param mileage
 	 */
-	public void setMileage(long mileage) {
+	public void setMileage(int mileage) {
 		this.mileage = mileage;
 	}
